chore: relocate tests
remove unused dependency
