+++ plugins/verdaccio-aws-s3-storage/tests/__fixtures__/pkg.ts
