+++ plugins/verdaccio-aws-s3-storage/tests/__mocks__/Config.ts
