+++ plugins/verdaccio-aws-s3-storage/tests/index.test.ts
@@ -1,11 +1,11 @@
 import { S3 } from 'aws-sdk';
 import { IPluginStorage } from '@verdaccio/types';
-import S3Database from '../';
+import S3Database from '../src/index';
 import Config from './__mocks__/Config';
 import logger from './__mocks__/Logger';
-import { deleteKeyPrefix } from '../deleteKeyPrefix';
-import { is404Error } from '../s3Errors';
-import { S3Config } from '../config';
+import { deleteKeyPrefix } from '../src/deleteKeyPrefix';
+import { is404Error } from '../src/s3Errors';
+import { S3Config } from '../src/config';
 
 describe('Local Database', () => {
   let db: IPluginStorage<S3Config>;
